+++ packages/contracts/contracts/core/plugin/IContractualMembership.sol
@@ -7,10 +7,10 @@ import {IDAO} from "../IDAO.sol";
 /// @notice An interface to be used by DAO plugins that define membership based on a contract. This can be
 /// - a token contract (e.g., (Governance-) ERC-20 and ERC-721) defining membership based on ownership / delegation.
 /// - a registry contract (e.g, `Addresslist`, the Proof of Humanity Registry) defining membership by curation of a list.
-interface IContractDefinedMembership {
+interface IContractualMembership {
     /// @notice Emitted to announce the membership being defined by a contract.
     /// @param definingContract The contract defining the membership.
-    event MembershipContractAnnounced(address indexed definingContract);
+    event ContractualMembershipAnnounced(address indexed definingContract);
 
     /// @notice Checks if an account is a member of the DAO.
     /// @param _account The address of the account to be checked.