+++ packages/contracts/contracts/voting/token/TokenVoting.sol
@@ -5,7 +5,7 @@ pragma solidity 0.8.10;
 import {IVotesUpgradeable} from "@openzeppelin/contracts-upgradeable/governance/utils/IVotesUpgradeable.sol";
 import {SafeCastUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/math/SafeCastUpgradeable.sol";
 
-import {IContractDefinedMembership} from "../../core/plugin/IContractDefinedMembership.sol";
+import {IContractualMembership} from "../../core/plugin/IContractualMembership.sol";
 import {IDAO} from "../../core/IDAO.sol";
 import {MajorityVotingBase} from "../majority/MajorityVotingBase.sol";
 import {IMajorityVoting} from "../majority/IMajorityVoting.sol";
@@ -14,7 +14,7 @@ import {IMajorityVoting} from "../majority/IMajorityVoting.sol";
 /// @author Aragon Association - 2021-2022
 /// @notice The majority voting implementation using an [OpenZepplin `Votes`](https://docs.openzeppelin.com/contracts/4.x/api/governance#Votes) compatible governance token.
 /// @dev This contract inherits from `MajorityVotingBase` and implements the `IMajorityVoting` interface.
-contract TokenVoting is IContractDefinedMembership, MajorityVotingBase {
+contract TokenVoting is IContractualMembership, MajorityVotingBase {
     using SafeCastUpgradeable for uint256;
 
     /// @notice The [ERC-165](https://eips.ethereum.org/EIPS/eip-165) interface ID of the contract.
@@ -41,7 +41,7 @@ contract TokenVoting is IContractDefinedMembership, MajorityVotingBase {
 
         votingToken = _token;
 
-        emit MembershipContractAnnounced({definingContract: address(_token)});
+        emit ContractualMembershipAnnounced({definingContract: address(_token)});
     }
 
     /// @notice Checks if this or the parent contract supports an interface by its ID.
@@ -113,7 +113,7 @@ contract TokenVoting is IContractDefinedMembership, MajorityVotingBase {
         }
     }
 
-    /// @inheritdoc IContractDefinedMembership
+    /// @inheritdoc IContractualMembership
     function isMember(address _account) external view returns (bool) {
         /// whatever condition
         return votingToken.getVotes(_account) >= getVotingSettings().minProposerVotingPower; //