+++ packages/subgraph/src/packages/multisig/multisig.ts
@@ -3,8 +3,8 @@ import {dataSource, store, BigInt} from '@graphprotocol/graph-ts';
 import {
   ProposalCreated,
   ProposalExecuted,
-  MembershipAnnounced,
-  MembershipRenounced,
+  IndividualMembershipAnnounced,
+  IndividualMembershipRenounced,
   Multisig,
   Approved,
   MultisigSettingsUpdated
@@ -156,7 +156,9 @@ export function handleProposalExecuted(event: ProposalExecuted): void {
   }
 }
 
-export function handleMembershipAnnounced(event: MembershipAnnounced): void {
+export function handleIndividualMembershipAnnounced(
+  event: IndividualMembershipAnnounced
+): void {
   const members = event.params.members;
   for (let index = 0; index < members.length; index++) {
     const member = members[index].toHexString();
@@ -173,7 +175,9 @@ export function handleMembershipAnnounced(event: MembershipAnnounced): void {
   }
 }
 
-export function handleMembershipRenounced(event: MembershipRenounced): void {
+export function handleIndividualMembershipRenounced(
+  event: IndividualMembershipRenounced
+): void {
   const members = event.params.members;
   for (let index = 0; index < members.length; index++) {
     const member = members[index].toHexString();